+++ src/db/schema/index.ts
@@ -28,3 +28,4 @@ export * from "./faqs";
 export * from "./newsletter";
 export * from "./blog-posts";
 export * from "./notifications";
+export * from "./payouts";